Fix type mismatch in LTI connections migration: use integer type for connectable_id
Agent-Logs-Url: https://github.com/mgmodell/CoLab/sessions/08d401d2-6ba3-43ac-bc45-8fdf54227213

+++ db/schema.rb
@@ -442,7 +442,7 @@ ActiveRecord::Schema[8.1].define(version: 2025_10_23_160856) do
 
   create_table "lti_connections", charset: "utf8mb3", collation: "utf8mb3_general_ci", force: :cascade do |t|
     t.string "connectable_type", null: false
-    t.bigint "connectable_id", null: false
+    t.integer "connectable_id", null: false
     t.string "line_item_url"
     t.string "ags_access_token_url"
     t.string "client_id"
